+++ src/main/java/com/appservice/all/Postgres/FacultyRepository.java
@@ -3,5 +3,5 @@ package com.appservice.all.Postgres;
 import com.appservice.all.Entities.Faculty;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FacultyRepository extends JpaRepository<Faculty, Integer> {
+public interface FacultyRepository extends JpaRepository<Faculty, Long> {
 }
